Adding validation and http status messages to course endpoint

+++ src/main/java/com/hyperskill/university/controllers/CourseController.java
@@ -1,5 +1,7 @@
 package com.hyperskill.university.controllers;
 
+import com.hyperskill.university.exceptions.DuplicatedException;
+import com.hyperskill.university.exceptions.InvalidException;
 import com.hyperskill.university.models.Course;
 import com.hyperskill.university.models.Student;
 import com.hyperskill.university.services.CourseService;
@@ -44,30 +46,69 @@ public class CourseController {
         }
         Course newCourse = courseService.addNewCourse(course);
         if (newCourse == null) {
-            return new ResponseEntity<>(Collections.singletonMap("msg", "Error creating course"), HttpStatus.CONFLICT);
+            return new ResponseEntity<>(Collections.singletonMap("msg", "Error creating course"), HttpStatus.INTERNAL_SERVER_ERROR);
         }
         return new ResponseEntity<>(newCourse, HttpStatus.CREATED) ;
     }
 
     @PutMapping(path = {"{courseId}"})
-    public void enrollStudent(@PathVariable("courseId") Integer courseId,
+    public ResponseEntity<Object> enrollStudent(@PathVariable("courseId") Integer courseId,
                               @RequestBody Student student) {
-        courseService.enrollStudent(courseId, student);
+        Course course = courseService.getCourseById(courseId);
+        if (course == null) {
+            String msg = "Course with ID: " + courseId + " not found";
+            return new ResponseEntity<>(Collections.singletonMap("msg", msg), HttpStatus.NOT_FOUND);
+        }
+        try {
+            Course updatedCourse = courseService.enrollStudent(courseId, student);
+            return new ResponseEntity<>(updatedCourse, HttpStatus.OK);
+        } catch (InvalidException invalidException) {
+            String msg = "Invalid student data";
+            return new ResponseEntity<>(Collections.singletonMap("msg", msg), HttpStatus.BAD_REQUEST);
+        } catch (DuplicatedException duplicatedException) {
+            String msg = "Student already enrolled";
+            return new ResponseEntity<>(Collections.singletonMap("msg", msg), HttpStatus.CONFLICT);
+        }
     }
 
     @PutMapping(path = {"/remove/{courseId}"})
-    public void unenrollStudent(@PathVariable("courseId") Integer courseId,
+    public ResponseEntity<Object> unenrollStudent(@PathVariable("courseId") Integer courseId,
                                 @RequestBody Student student) {
-        courseService.removeStudentFromCourse(courseId, student);
+        Course course = courseService.getCourseById(courseId);
+        if (course == null) {
+            String msg = "Course with ID: " + courseId + " not found";
+            return new ResponseEntity<>(Collections.singletonMap("msg", msg), HttpStatus.NOT_FOUND);
+        }
+        try {
+            Course updatedCourse = courseService.removeStudentFromCourse(courseId, student);
+            return new ResponseEntity<>(updatedCourse, HttpStatus.OK);
+        } catch (InvalidException invalidException) {
+            String msg = "Invalid student data";
+            return new ResponseEntity<>(Collections.singletonMap("msg", msg), HttpStatus.BAD_REQUEST);
+        } catch (DuplicatedException duplicatedException) {
+            String msg = "Student is not enrolled";
+            return new ResponseEntity<>(Collections.singletonMap("msg", msg), HttpStatus.CONFLICT);
+        }
     }
 
     @DeleteMapping(path = {"{courseId}"})
-    public void deleteCourse(@PathVariable("courseId") Integer courseId) {
+    public ResponseEntity<Void> deleteCourse(@PathVariable("courseId") Integer courseId) {
+        Course course = courseService.getCourseById(courseId);
+        if (course == null) {
+            String msg = "Course with ID: " + courseId + " not found";
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         courseService.removeCourse(courseId);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @GetMapping(path = {"/dep/{departmentId}"})
-    public List<Course> getCoursesByDepartmentId(@PathVariable("departmentId") Integer departmentId) {
-        return courseService.getCoursesByDepartment(departmentId);
+    public ResponseEntity<Object> getCoursesByDepartmentId(@PathVariable("departmentId") Integer departmentId) {
+        List<Course> courses = courseService.getCoursesByDepartment(departmentId);
+        if (courses == null) {
+            String msg = "Department with ID: " + departmentId + " not found";
+            return new ResponseEntity<>(Collections.singletonMap("msg", msg), HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(courses, HttpStatus.OK);
     }
 }
